inactivating same state as entering
Discovered an issue where, if by some chance, the states are such that you could possibly enter a state while at the same time inactivating it, then of course the state is not properly entered. This happened to me bc I was switching from a state, which was then inactivated, but a modal from that inactive state was never destroyed or hidden. Using this modal I could attempt to switch to the same state again, but under different params. This caused the state the modal originated from to be reactivated, the target state to be inactivated, and then, paradoxically, the target state to be entered! This whole issue is more of a state-setup issue but nevertheless I added code to find this scenario and to exit the target state and remove the inactivation of the target state. This causes the new state to be exited and entered anew, with its new params, which is the expected behavior.